requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.43
idna==3.7
multidict==6.0.5
py-cord==2.5.0
smmap==5.0.1
typing_extensions==4.11.0
yarl==1.9.4
